Showing best solutions and consensus level. Changed login view.

+++ tdm.domain/src/main/java/com/tdm/domain/model/problem/dto/CurrentConsensus.java
@@ -6,7 +6,7 @@ import com.tdm.domain.model.idea.dto.SolutionIdea;
 
 public interface CurrentConsensus {
 
-	double getConsensusLevel();
+	int getConsensusLevel();
 
 	String getProblemId();
